# swagger.yml
openapi: 3.0.3
info:
  title: ObpPay API
  description: API officielle d’ObpPay – Auth, Wallet, Transactions & Loan system.
  version: 1.0.0

servers:
  - url: http://127.0.0.1:8000/api
    description: Serveur local de développement

paths:

  /auth/register:
    post:
      summary: Start or complete registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                phone: { type: string }
                password: { type: string }
                otp: { type: string }
                obp_id: { type: string }
      responses:
        '200':
          description: Registration step started
        '201':
          description: Registration completed
        '422':
          description: Validation error

  /auth/login:
    post:
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                phone: { type: string }
                password: { type: string }
      responses:
        '200':
          description: Successful login
        '401':
          description: Invalid credentials

  /auth/me:
    get:
      summary: Get authenticated user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile with wallet

  /auth/logout:
    post:
      summary: Logout user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful logout

  /auth/send-otp:
    post:
      summary: Send OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                phone: { type: string }
      responses:
        '200':
          description: OTP sent

  /auth/verify-otp:
    post:
      summary: Verify OTP code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                phone: { type: string }
                code: { type: string }
      responses:
        '200':
          description: OTP verified

  /auth/profile:
    put:
      summary: Update user profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                phone: { type: string }
                email: { type: string }
      responses:
        '200':
          description: Profile updated
        '422':
          description: Validation error

  /auth/change-password:
    put:
      summary: Change user password
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                current_password: { type: string }
                new_password: { type: string }
      responses:
        '200':
          description: Password changed
        '422':
          description: Validation error

  /user/update-photo:
    post:
      summary: Update user photo
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                photo:
                  type: string
                  format: binary
      responses:
        '200':
          description: Photo updated
        '422':
          description: Validation error

  /user/by-obp/{obp_id}:
    get:
      summary: Fetch user by OBP ID
      parameters:
        - in: path
          name: obp_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User found
        '404':
          description: User not found

  /user/score/latest:
    get:
      summary: Get latest user score
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Latest score
        '404':
          description: No score found

  /user/score:
    get:
      summary: Get all user scores
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of scores
        '404':
          description: No scores found

  /wallet/balance:
    get:
      summary: Get wallet balance
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current balance

  /wallet/transactions:
    get:
      summary: List user transactions
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of transactions

  /wallet/deposit:
    post:
      summary: Deposit money
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount: { type: number }
      responses:
        '200':
          description: Deposit successful

  /wallet/transfer:
    post:
      summary: Transfer funds
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                to: { type: string }
                amount: { type: number }
      responses:
        '200':
          description: Transfer successful

  /loan/eligibility:
    get:
      summary: Check loan eligibility
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Eligibility info

  /loan/request:
    post:
      summary: Request a loan
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount: { type: number }
                duration: { type: number }
      responses:
        '200':
          description: Loan request submitted

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT